update contract to split execute and query (#240)
